Added improved master indicators module logic
