+++ sqlMesh/models/master/indicators.py
@@ -3,6 +3,7 @@ import os
 from sqlmesh.core.macros import MacroEvaluator
 from sqlmesh.core.model import model
 from macros.ibis_expressions import generate_ibis_table
+from macros.utils import find_indicator_models
 from constants import SQLMESH_DIR
 
 COLUMN_SCHEMA = {
@@ -14,22 +15,6 @@ COLUMN_SCHEMA = {
     "qualifier": "String",
     "indicator_description": "String",
 }
-
-
-def find_indicator_models():
-    """Find all models ending with _indicators in the sources directory."""
-    indicator_models = []
-    sources_dir = os.path.join(SQLMESH_DIR, "models", "sources")
-    
-    for source in os.listdir(sources_dir):
-        source_dir = os.path.join(sources_dir, source)
-        if os.path.isdir(source_dir):
-            for file in os.listdir(source_dir):
-                if file.endswith('_indicators.py'):
-                    module_name = f"models.sources.{source}.{file[:-3]}"
-                    indicator_models.append((source, module_name))
-    
-    return indicator_models
 
 
 @model(
@@ -45,18 +30,23 @@ def entrypoint(evaluator: MacroEvaluator) -> str:
     # Import each model and get its table
     tables = []
     for source, module_name in indicator_models:
-        # Dynamically import the module
-        module = __import__(module_name, fromlist=['COLUMN_SCHEMA'])
-        
-        # Generate table for this source
-        table = generate_ibis_table(
-            evaluator,
-            table_name=source,
-            schema_name="sources",
-            column_schema=module.COLUMN_SCHEMA,
-        )
-        # Add source column
-        tables.append(table.mutate(source=ibis.literal(source)))
+        try:
+            # Dynamically import the module
+            module = __import__(module_name, fromlist=['COLUMN_SCHEMA'])
+            
+            # Generate table for this source
+            table = generate_ibis_table(
+                evaluator,
+                table_name=source,
+                schema_name="sources",
+                column_schema=module.COLUMN_SCHEMA,
+            )
+            # Add source column
+            tables.append(table.mutate(source=ibis.literal(source)))
+        except ImportError:
+            raise ImportError(f"Could not import module: {module_name}")
+        except AttributeError:
+            raise AttributeError(f"Module {module_name} does not have COLUMN_SCHEMA")
     
     # Union all tables
     if not tables:
